add files of processed data and the original code
three data files are to answer the question of "which area(census tract) receives how many loans". loan_amount is the total data for the loan originated in suffolk county. avg_loan_each_census_tract gives the average loan for census tracts. avg_loan_suffolk_county is for the average loan for the entire Suffolk county